// x-pack/plugins/security_solution/public/common/components/markdown_editor/types.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License;
 * you may not use this file except in compliance with the Elastic License.
 */

export interface CursorPosition {
  start: number;
  end: number;
}